Added a base class for plot configurations
