+++ python/__init__.py
@@ -31,7 +31,7 @@ except ImportError:
 	pass
 
 # import any pure python here
+from bokeh_plot_config import bokeh_plot_config
 from time_sink_f import time_sink_f
 from time_sink_c import time_sink_c
-
 #
